SecondGlobalCommitPostTest

+++ SYACTest/Entitys/Products.cs
@@ -11,7 +11,7 @@ namespace SYACTest.Entitys
         public List<OrderProducts> OrderProducts { get; set; }
         public string productCode { get; set; }
         public string productname { get; set; }
-        public int unitValue { get; set; }
+        public decimal unitValue { get; set; }
 
     }
 }
